improve RecordReader API design : make the getTotalRecords method return Long (instead of long) to be able to return null if the total records number cannot be calculated in advance

+++ easybatch-jdbc/src/main/java/io/github/benas/easybatch/jdbc/JdbcRecordReader.java
@@ -95,8 +95,8 @@ public class JdbcRecordReader implements RecordReader {
     }
 
     @Override
-    public long getTotalRecords() {
-        int rowCount = 0;
+    public Long getTotalRecords() {
+        long rowCount = 0;
 
         try {
             if (resultSet.last()) {
